+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/models/AgeResolution.java
@@ -51,6 +51,11 @@ public final class AgeResolution extends BaseResolution {
         return this.value;
     }
 
+    @Override
+    public ResolutionKind getType() {
+        return ResolutionKind.AGE_RESOLUTION;
+    }
+
     private void setUnit(AgeUnit unit) {
         this.unit = unit;
     }