+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/models/VolumeResolution.java
@@ -51,6 +51,11 @@ public final class VolumeResolution extends BaseResolution {
         return this.value;
     }
 
+    @Override
+    public ResolutionKind getType() {
+        return ResolutionKind.VOLUME_RESOLUTION;
+    }
+
     private void setUnit(VolumeUnit unit) {
         this.unit = unit;
     }