+++ sdk/textanalytics/azure-ai-textanalytics/src/samples/java/com/azure/ai/textanalytics/batch/RecognizeEntitiesBatchDocuments.java
@@ -6,7 +6,6 @@ package com.azure.ai.textanalytics.batch;
 import com.azure.ai.textanalytics.TextAnalyticsClient;
 import com.azure.ai.textanalytics.TextAnalyticsClientBuilder;
 import com.azure.ai.textanalytics.models.AgeResolution;
-import com.azure.ai.textanalytics.models.BaseResolution;
 import com.azure.ai.textanalytics.models.RecognizeEntitiesResult;
 import com.azure.ai.textanalytics.models.TextAnalyticsRequestOptions;
 import com.azure.ai.textanalytics.models.TextDocumentBatchStatistics;
@@ -15,6 +14,7 @@ import com.azure.ai.textanalytics.models.WeightResolution;
 import com.azure.ai.textanalytics.util.RecognizeEntitiesResultCollection;
 import com.azure.core.credential.AzureKeyCredential;
 import com.azure.core.http.rest.Response;
+import com.azure.core.util.Configuration;
 import com.azure.core.util.Context;
 
 import java.util.Arrays;
@@ -33,8 +33,8 @@ public class RecognizeEntitiesBatchDocuments {
     public static void main(String[] args) {
         // Instantiate a client that will be used to call the service.
         TextAnalyticsClient client = new TextAnalyticsClientBuilder()
-            .credential(new AzureKeyCredential("{key}"))
-            .endpoint("{endpoint}")
+            .credential(new AzureKeyCredential(Configuration.getGlobalConfiguration().get("AZURE_TEXT_ANALYTICS_API_KEY")))
+            .endpoint(Configuration.getGlobalConfiguration().get("AZURE_TEXT_ANALYTICS_ENDPOINT"))
             .buildClient();
 
         // The texts that need be analyzed.
@@ -76,20 +76,23 @@ public class RecognizeEntitiesBatchDocuments {
                     System.out.printf(
                             "Recognized entity: %s, entity category: %s, entity subcategory: %s, confidence score: %f.%n",
                             entity.getText(), entity.getCategory(), entity.getSubcategory(), entity.getConfidenceScore());
-                    Iterable<? extends BaseResolution> resolutions = entity.getResolutions();
-                    if (resolutions != null) {
-                        for (BaseResolution resolution : resolutions) {
-                            if (resolution instanceof WeightResolution) {
+                    entity.getResolutions().forEach(resolution -> {
+                        switch (resolution.getType()) {
+                            case AGE_RESOLUTION:
+                                AgeResolution ageResolution = (AgeResolution) resolution;
+                                System.out.printf("\tAgeResolution: unit: %s. value: %f.%n", ageResolution.getUnit(),
+                                    ageResolution.getValue());
+                                break;
+                            case WEIGHT_RESOLUTION:
                                 WeightResolution weightResolution = (WeightResolution) resolution;
                                 System.out.printf("\tWeightResolution: unit: %s. value: %f.%n", weightResolution.getUnit(),
-                                        weightResolution.getValue());
-                            } else if (resolution instanceof AgeResolution) {
-                                AgeResolution weightResolution = (AgeResolution) resolution;
-                                System.out.printf("\tAgeResolution: unit: %s. value: %f.%n", weightResolution.getUnit(),
-                                        weightResolution.getValue());
-                            }
+                                    weightResolution.getValue());
+                                break;
+                            default:
+                                System.out.printf("\tUnmatched resolution type: %s", resolution.getType());
+                                break;
                         }
-                    }
+                    });
                 });
             }
         }