+++ sdk/textanalytics/azure-ai-textanalytics/src/samples/java/com/azure/ai/textanalytics/batch/RecognizeEntitiesBatchDocumentsAsync.java
@@ -6,7 +6,6 @@ package com.azure.ai.textanalytics.batch;
 import com.azure.ai.textanalytics.TextAnalyticsAsyncClient;
 import com.azure.ai.textanalytics.TextAnalyticsClientBuilder;
 import com.azure.ai.textanalytics.models.AgeResolution;
-import com.azure.ai.textanalytics.models.BaseResolution;
 import com.azure.ai.textanalytics.models.RecognizeEntitiesResult;
 import com.azure.ai.textanalytics.models.TextAnalyticsRequestOptions;
 import com.azure.ai.textanalytics.models.TextDocumentBatchStatistics;
@@ -14,6 +13,7 @@ import com.azure.ai.textanalytics.models.TextDocumentInput;
 import com.azure.ai.textanalytics.models.WeightResolution;
 import com.azure.ai.textanalytics.util.RecognizeEntitiesResultCollection;
 import com.azure.core.credential.AzureKeyCredential;
+import com.azure.core.util.Configuration;
 
 import java.util.Arrays;
 import java.util.List;
@@ -32,8 +32,8 @@ public class RecognizeEntitiesBatchDocumentsAsync {
     public static void main(String[] args) {
         // Instantiate a client that will be used to call the service.
         TextAnalyticsAsyncClient client = new TextAnalyticsClientBuilder()
-            .credential(new AzureKeyCredential("{key}"))
-            .endpoint("{endpoint}")
+            .credential(new AzureKeyCredential(Configuration.getGlobalConfiguration().get("AZURE_TEXT_ANALYTICS_API_KEY")))
+            .endpoint(Configuration.getGlobalConfiguration().get("AZURE_TEXT_ANALYTICS_ENDPOINT"))
             .buildAsyncClient();
 
         // The texts that need be analyzed.
@@ -42,8 +42,9 @@ public class RecognizeEntitiesBatchDocumentsAsync {
             new TextDocumentInput("B", "The cat is 1 year old and weighs 10 pounds.").setLanguage("en")
         );
 
-        TextAnalyticsRequestOptions requestOptions = new TextAnalyticsRequestOptions().setIncludeStatistics(true)
-                .setModelVersion("2022-10-01-preview");
+        TextAnalyticsRequestOptions requestOptions = new TextAnalyticsRequestOptions()
+            .setIncludeStatistics(true)
+            .setModelVersion("2022-10-01-preview");
 
         // Recognizing entities for each document in a batch of documents
         client.recognizeEntitiesBatchWithResponse(documents, requestOptions).subscribe(
@@ -73,20 +74,22 @@ public class RecognizeEntitiesBatchDocumentsAsync {
                             System.out.printf(
                                     "Recognized entity: %s, entity category: %s, entity subcategory: %s, confidence score: %f.%n",
                                     entity.getText(), entity.getCategory(), entity.getSubcategory(), entity.getConfidenceScore());
-                            Iterable<? extends BaseResolution> resolutions = entity.getResolutions();
-                            if (resolutions != null) {
-                                for (BaseResolution resolution : resolutions) {
-                                    if (resolution instanceof WeightResolution) {
+                            entity.getResolutions().forEach(resolution -> {
+                                switch (resolution.getType()) {
+                                    case AGE_RESOLUTION:
+                                        AgeResolution ageResolution = (AgeResolution) resolution;
+                                        System.out.printf("\tAgeResolution: unit: %s. value: %f.%n", ageResolution.getUnit(),
+                                            ageResolution.getValue());
+                                        break;
+                                    case WEIGHT_RESOLUTION:
                                         WeightResolution weightResolution = (WeightResolution) resolution;
                                         System.out.printf("\tWeightResolution: unit: %s. value: %f.%n", weightResolution.getUnit(),
-                                                weightResolution.getValue());
-                                    } else if (resolution instanceof AgeResolution) {
-                                        AgeResolution weightResolution = (AgeResolution) resolution;
-                                        System.out.printf("\tAgeResolution: unit: %s. value: %f.%n", weightResolution.getUnit(),
-                                                weightResolution.getValue());
-                                    }
+                                            weightResolution.getValue());
+                                        break;
+                                    default:
+                                        break;
                                 }
-                            }
+                            });
                         });
                     }
                 }